Entry/exit times are now able to be logged in the database
When a student has their tag read by an RFID reader, their status in the
database is now updated regardless of the time the tag was read. Further
checks are now made to determine whether a student has arrived late or
departed early. Parents are now notified if either of these criteria are
met via email and are asked to log into the web application to provide
an explanation.